ms_web/xs-app/xs-web/src/main/java/com/toney/mvc/web/controller/HomeController.java
package com.toney.mvc.web.controller;

import java.text.DateFormat;
import java.util.Date;
import java.util.Locale;

import org.apache.commons.lang.StringEscapeUtils;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

/**
 * Handles requests for the application home page.
 *  首页
 */
@Controller
public class HomeController {

    private static final Logger LOGGER = LoggerFactory.getLogger(HomeController.class);

    /**
     * Simply selects the home view to render by returning its name.
     */
    @RequestMapping(value = "/index", method = RequestMethod.GET)
    public String home(Locale locale, Model model) {
        LOGGER.info("Welcome home! the client locale is " + locale.toString());

        Date date = new Date();
        DateFormat dateFormat = DateFormat.getDateTimeInstance(DateFormat.LONG, DateFormat.LONG, locale);

        String formattedDate = dateFormat.format(date);

        model.addAttribute("serverTime", formattedDate);
        String xssValue = "<script></script>";
        xssValue = StringEscapeUtils.escapeHtml(xssValue);
        xssValue = StringEscapeUtils.escapeJavaScript(xssValue);
        model.addAttribute("xssValue", xssValue);

        return "home";
    }
}
